Breaking changes: 1. Make ConvexAppBar as a standlone Widget; 2. Remove dependency of FAB & Location;
