+++ example/lib/custom_appbar_sample.dart
@@ -10,12 +10,9 @@ class CustomAppBarDemo extends StatefulWidget {
 }
 
 class _State extends State<CustomAppBarDemo> {
-  int _currentIndex = 0;
-  static const INDEX_PUBLISH = 2;
-  List<TabItem> _navigationItems = <TabItem>[
+  List<TabItem> items = <TabItem>[
     TabItem(icon: Icons.home, title: 'Home'),
     TabItem(icon: Icons.map, title: 'Discovery'),
-    TabItem(title: ''),
     TabItem(icon: Icons.message, title: 'Message'),
     TabItem(icon: Icons.people, title: 'Profile')
   ];
@@ -44,26 +41,53 @@ class _State extends State<CustomAppBarDemo> {
 
   @override
   Widget build(BuildContext context) {
-    var convexColor =
-        _currentIndex == INDEX_PUBLISH ? Colors.white : Colors.white60;
     return Scaffold(
       appBar: AppBar(title: const Text('Custom ConvexAppBar')),
       body: paletteBody(),
-      floatingActionButton: GestureDetector(
-        onTap: () => _onItemTapped(INDEX_PUBLISH),
-        child: fabContent(convexColor),
-      ),
-      floatingActionButtonLocation: ConvexAppBar.centerDocked,
       bottomNavigationBar: ConvexAppBar.builder(
-          count: 5,
-          backgroundColor: _tabBackgroundColor,
-          builder: (BuildContext context, int index) {
-            var data = _navigationItems[index];
-            var color = _currentIndex == index ? Colors.white : Colors.white60;
-            return GestureDetector(
-                onTap: () => _onItemTapped(index),
-                child: tabContent(data, color));
-          }),
+        count: items.length,
+        backgroundColor: _tabBackgroundColor,
+        tabBuilder: (BuildContext context, int index, bool active) {
+          var navigationItem = items[index];
+          var _color = active ? Colors.white : Colors.white60;
+          var _icon = active
+              ? navigationItem.activeIcon ?? navigationItem.icon
+              : navigationItem.icon;
+          return Container(
+            color: Colors.transparent,
+            padding: EdgeInsets.only(bottom: 2),
+            child: Column(
+              mainAxisAlignment: MainAxisAlignment.end,
+              children: <Widget>[
+                Icon(_icon, color: _color),
+                Text(navigationItem.title, style: TextStyle(color: _color))
+              ],
+            ),
+          );
+        },
+        actionBuilder: (BuildContext context, int index, bool active) {
+          var _color = active ? Colors.white : Colors.white60;
+
+          return Stack(
+            alignment: Alignment.center,
+            children: <Widget>[
+              SizedBox(
+                width: 60,
+                height: 60,
+                child: Container(
+                  decoration:
+                      BoxDecoration(shape: BoxShape.circle, color: _color),
+                  child: Icon(
+                    Icons.add,
+                    size: 40,
+                    color: _tabBackgroundColor,
+                  ),
+                ),
+              )
+            ],
+          );
+        },
+      ),
     );
   }
 
@@ -80,21 +104,6 @@ class _State extends State<CustomAppBarDemo> {
         ));
   }
 
-  Container fabContent(Color convexColor) {
-    return Container(
-      width: 60,
-      height: 80,
-      padding: EdgeInsets.only(bottom: 2),
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.end,
-        children: <Widget>[
-          Icon(Icons.add, size: 50, color: convexColor),
-          Text('Publish', style: TextStyle(color: convexColor)),
-        ],
-      ),
-    );
-  }
-
   GridView paletteBody() {
     return GridView.count(
       crossAxisCount: 5,
@@ -108,11 +117,6 @@ class _State extends State<CustomAppBarDemo> {
               ))
           .toList(),
     );
-  }
-
-  bool _onItemTapped(int index) {
-    setState(() => _currentIndex = index);
-    return true;
   }
 
   void _onColorChanged(Color color) {
