+++ lib/convex_bottom_bar.dart
@@ -1,4 +1,5 @@
 library bulged_bar;
 
-export 'src/bottom_navigation_bar.dart';
+export 'src/experiment/deprecated.dart';
 export 'src/item.dart';
+export 'src/bar.dart';
