+++ lib/src/experiment/extend_location.dart
@@ -1,0 +1,55 @@
+import 'dart:ui';
+import 'dart:math' as math;
+
+import 'package:flutter/material.dart';
+
+class ExtendLocation extends FloatingActionButtonLocation {
+  final bool snackBarCompact;
+  final bool bottomSheetCompact;
+  final double verticalOffset;
+
+  const ExtendLocation(
+      {this.snackBarCompact = false,
+      this.bottomSheetCompact = false,
+      this.verticalOffset: 10});
+
+  @override
+  Offset getOffset(ScaffoldPrelayoutGeometry scaffoldGeometry) {
+    final double fabX = (scaffoldGeometry.scaffoldSize.width -
+            scaffoldGeometry.floatingActionButtonSize.width) /
+        2.0;
+    return Offset(fabX, getDockedY(scaffoldGeometry) + verticalOffset);
+  }
+
+  @override
+  String toString() => 'FloatingActionButtonLocation.centerDocked';
+
+  // Positions the Y coordinate of the [FloatingActionButton] at a height
+  // where it docks to the [BottomAppBar].
+  @protected
+  double getDockedY(ScaffoldPrelayoutGeometry scaffoldGeometry) {
+    final double contentBottom = scaffoldGeometry.contentBottom;
+    final double bottomSheetHeight = scaffoldGeometry.bottomSheetSize.height;
+    final double fabHeight = scaffoldGeometry.floatingActionButtonSize.height;
+    final double snackBarHeight = scaffoldGeometry.snackBarSize.height;
+
+    double fabY = contentBottom - fabHeight / 2.0;
+    // The FAB should sit with a margin between it and the snack bar.
+    if (snackBarCompact && snackBarHeight > 0.0)
+      fabY = math.min(
+          fabY,
+          contentBottom -
+              snackBarHeight -
+              fabHeight -
+              kFloatingActionButtonMargin);
+    // The FAB should sit with its center in front of the top of the bottom sheet.
+    if (bottomSheetCompact && bottomSheetHeight > 0.0)
+      fabY =
+          math.min(fabY, contentBottom - bottomSheetHeight - fabHeight / 2.0);
+
+    final double maxFabY = scaffoldGeometry.scaffoldSize.height - fabHeight;
+    return math.min(maxFabY, fabY);
+  }
+
+  static const ExtendLocation centerDocked = ExtendLocation();
+}
